api_service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
httpx==0.19.0
grpcio==1.40.0
grpcio-tools==1.40.0
protobuf==3.17.3
python-jose[cryptography]==3.3.0
pydantic==1.8.2
requests==2.26.0
pytest==6.2.5
